+++ src/screens/WelcomeScreen/ImportWalletModal/ImportWalletModal.tsx
@@ -1,35 +1,19 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import { SafeAreaView, View } from 'react-native';
 import { Text, Button, TextArea, ModalHeader, LoadingScreen } from '@components';
 import KeyboardSpacer from 'react-native-keyboard-spacer';
-import { useFormProgress, useNavigation } from '@hooks';
 import { ImportWalletModalProps } from './ImportWalletModal.types';
 import { useImportWalletModal } from './useImportWallet.hooks';
 import SelectImportMethodModal from './SelectImportMethodModal/SelectImportMethodModal';
 import { styles } from './ImportWalletModal.styles';
 
 const ImportWalletModal: React.FC<ImportWalletModalProps> = ({ onImportFinished, onDismiss, visible }) => {
-	const { currentStep, reset, goForward, goBack } = useFormProgress();
-	const { text, setText, importing, onImportWallet } = useImportWalletModal({ onImportFinished });
-	const navigation = useNavigation();
-
-	const onICloudBackup = () => {
-		navigation.navigate('BackupToICloudScreen', { missingPassword: false, restoreBackups: true });
-	};
-
-	const onBack = () => {
-		if (currentStep === 0) {
-			onDismiss();
-		} else {
-			goBack();
-		}
-	};
-
-	useEffect(() => {
-		if (!visible) {
-			reset();
-		}
-	}, [visible]);
+	const { text, setText, importing, onImportWallet, onBack, currentStep, goForward, onICloudBackup } =
+		useImportWalletModal({
+			onImportFinished,
+			onDismiss,
+			visible
+		});
 
 	return (
 		<SafeAreaView>
